requirements.txt
aiogram>=2.10.1
envparse>=0.2.0
APScheduler>=3.6.3
pytz>=2020.1
gino>=1.0.1
alembic>=1.4.3
SQLAlchemy>=1.3.19
redis>=3.5.3
aioredis>=1.3.1
Pillow>=8.0.1
httpx>=0.16.1